factor out render_phase.dart
