+++ lib/chronosgl.dart
@@ -30,67 +30,9 @@ Node UnitNode( WEBGL.RenderingContext gl) {
   return new Node("unit-mesh", UnitQuad, EmptyMaterial);
 }
 
-/// RenderPhase represents a sequence of RenderPrograms.
-class RenderPhase extends NamedEntity {
-  final WEBGL.RenderingContext _gl;
-  ChronosFramebuffer _framebuffer;
-  final List<RenderProgram> _programs = [];
-  //final VM.Matrix4 _pMatrix = new VM.Matrix4.identity();
-  bool clearColorBuffer = true;
-  bool clearDepthBuffer = true;
-  int viewPortX = 0;
-  int viewPortY = 0;
-  int viewPortW = 0;
-  int viewPortH = 0;
 
-  RenderPhase(String name, this._gl, [this._framebuffer = null]) : super(name);
-
-  void SetFramebuffer(ChronosFramebuffer fb) {
-    _framebuffer = fb;
-  }
-
-  void draw(List<RenderInputProvider> inputs, [List<DrawStats> stats = null]) {
-    if (_framebuffer == null) {
-      _gl.bindFramebuffer(WEBGL.FRAMEBUFFER, null);
-    } else {
-      _gl.bindFramebuffer(WEBGL.FRAMEBUFFER, _framebuffer.framebuffer);
-    }
-    assert(viewPortW > 0 && viewPortH > 0);
-    _gl.viewport(viewPortX, viewPortY, viewPortW, viewPortH);
-
-    if (clearColorBuffer || clearDepthBuffer) {
-      int mode = 0;
-      if (clearColorBuffer) mode |= WEBGL.COLOR_BUFFER_BIT;
-      if (clearDepthBuffer) mode |= WEBGL.DEPTH_BUFFER_BIT;
-      _gl.clear(mode);
-    }
-
-    for (RenderProgram prg in _programs) {
-      if (!prg.enabled) continue;
-      for (RenderInputProvider p in inputs) {
-        p.AddRenderInputs(prg);
-      }
-      prg.draw(stats);
-      for (RenderInputProvider p in inputs) {
-        p.RemoveRenderInputs(prg);
-      }
-    }
-  }
-
-  void AddRenderProgram(RenderProgram s) {
-    _programs.add(s);
-  }
-
-  RenderProgram createProgram(List<ShaderObject> so) {
-    ShaderProgram prg = new ShaderProgram(so[0].name, _gl, so[0], so[1]);
-    RenderProgram pn = new RenderProgram(so[0].name, prg);
-    AddRenderProgram(pn);
-    return pn;
-  }
-}
 
 class ChronosGL {
-  static bool useElementIndexUint = false;
   var elementIndexUintExt;
 
   WEBGL.RenderingContext gl;
@@ -121,7 +63,7 @@ class ChronosGL {
       if (elementIndexUintExt == null) {
         throw "Error: OES_element_index_uint is not supported";
       }
-      ChronosGL.useElementIndexUint = useElementIndexUint;
+      globalUseElementIndexUint = useElementIndexUint;
     }
 
     //print( gl.getSupportedExtensions());
